+++ examples/spy.py
@@ -13,7 +13,7 @@ class Plugin(object):
         self.context = context
         self.channel = context.config.channel
         try:
-            self.chater = context.config.chater
+            self.chater = context.botnet['bot_chater']
         except Exception:
             self.chater = None
 
@@ -27,28 +27,3 @@ class Plugin(object):
         if self.chater:
             self.chater.privmsg(self.chater.config.channel,
                                 '{0}: {1}'.format(mask.nick, data))
-
-
-def main():
-    loop = asyncio.get_event_loop()
-    config = {
-        'loop': loop,
-        'nick': 'chater',
-        'channel': '#irc3',
-        'host': 'irc.freenode.net',
-        'debug': True,
-        'verbose': True,
-        'raw': True,
-        'includes': ['spy'],
-    }
-    chater = irc3.IrcBot.from_config(config)
-    chater.run(forever=False)
-    config.update({
-        'nick': 'spyer',
-        'channel': '#irc3_dev',
-        'chater': chater,
-    })
-    spyer = irc3.IrcBot.from_config(config)
-    spyer.run(forever=False)
-    print('serving')
-    loop.run_forever()